@@ -15,52 +15,53 @@
     "build-stories": "build-storybook --no-dll"
   },
   "dependencies": {
-    "@theme-ui/color": "^0.3.1",
-    "@theme-ui/prism": "^0.3.0",
-    "babel-plugin-styled-components": "^1.10.7",
-    "fast-glob": "^3.2.4",
-    "fuse.js": "^6.4.1",
+    "@emotion/react": "^11.9.0",
+    "@theme-ui/color": "^0.14.5",
+    "@theme-ui/prism": "^0.14.5",
+    "babel-plugin-styled-components": "^2.0.7",
+    "fast-glob": "^3.2.11",
+    "fuse.js": "^6.6.2",
     "gray-matter": "^4.0.2",
     "lodash": "^4.17.20",
-    "markdown-link-extractor": "^2.0.1",
-    "mdast-util-toc": "^5.1.0",
-    "next": "^12.0.0",
-    "next-mdx-relations": "^0.0.13",
+    "markdown-link-extractor": "^4.0.1",
+    "mdast-util-toc": "^6.1.0",
+    "next": "^12.1.6",
+    "next-mdx-relations": "^0.0.14",
     "next-mdx-remote": "^4.0.3",
-    "prism-react-renderer": "^1.1.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "remark-autolink-headings": "^6.0.1",
-    "remark-slug": "^6.0.0",
-    "styled-components": "^5.1.1",
+    "prism-react-renderer": "^1.3.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "remark-autolink-headings": "^7.0.1",
+    "remark-slug": "^7.0.1",
+    "styled-components": "^5.3.5",
     "styled-system": "^5.1.5",
-    "theme-ui": "^0.3.1",
-    "unist-util-visit": "^2.0.3"
+    "theme-ui": "^0.14.5",
+    "unist-util-visit": "^4.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.6",
-    "@babel/plugin-syntax-jsx": "^7.16.0",
-    "@storybook/addon-actions": "^6.0.28",
-    "@storybook/addon-essentials": "^6.0.28",
-    "@storybook/addon-links": "^6.0.28",
-    "@storybook/react": "^6.0.28",
+    "@babel/core": "^7.17.10",
+    "@babel/plugin-syntax-jsx": "^7.16.7",
+    "@storybook/addon-actions": "^6.4.22",
+    "@storybook/addon-essentials": "^6.4.22",
+    "@storybook/addon-links": "^6.4.22",
+    "@storybook/react": "^6.4.22",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.1.0",
-    "babel-loader": "^8.1.0",
+    "babel-jest": "^28.1.0",
+    "babel-loader": "^8.2.5",
     "babel-plugin-module-resolver": "^4.0.0",
-    "eslint": "^7.5.0",
-    "eslint-config-airbnb": "^18.1.0",
-    "eslint-config-prettier": "^6.11.0",
+    "eslint": "^8.15.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-import-resolver-babel-module": "^5.1.2",
-    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-react": "^7.20.0",
-    "jest": "^26.6.3",
-    "prettier": "^2.0.5",
-    "react-is": "^17.0.1",
-    "react-test-renderer": "^16.13.1"
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.29.4",
+    "jest": "^28.1.0",
+    "prettier": "^2.6.2",
+    "react-is": "^18.1.0",
+    "react-test-renderer": "^18.1.0"
   }
 }
