@@ -20,5 +20,7 @@ module.exports = {
   globals: {
     __PATH_PREFIX__: ``
   },
-  testURL: `http://localhost`
+  testEnvironmentOptions: {
+    url: `http://localhost`
+  }
 };
